requirements.txt
psycopg2-binary==2.8.5
requests==2.24.0
ujson==3.1.0
kafka-python==2.0.1
click==7.1.2